Pivot data from wide to long

pivot_longer() "lengthens" data, increasing the number of rows and decreasing the number of columns. The inverse transformation is pivot_wider().

pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  ...
)

Arguments

data

data.frame. The data to pivot.

cols

<poor-select>. Columns to pivot into longer format.

names_to

character(n). The name of the new column(s) that will contain the column names.

names_prefix

character(1). A regular expression used to remove matching text from the start of each variable name.

names_sep, names_pattern

character(1). If names_to contains multiple values, this argument controls how the column name is broken up. names_pattern takes a regular expression containing matching groups (()⁠).

values_to

character(n). The name of the new column(s) that will contain the values of the pivoted variables.

values_drop_na

logical(1). If TRUE, will drop rows that contain only NA in the values_to column. This effectively converts explicit missing values to implicit missing values, and should generally be used only when missing values in data were created by its structure.

...

Additional arguments passed on to methods.

Value

A data.frame.

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))
# Customizing the names
pivot_longer(
  data = wide_data,
  cols = c(1, 2),
  names_to = "Column",
  values_to = "Numbers"
)
#>             X3         X4          X5 Column    Numbers
#> 1  -0.05355055 -0.5975271 -0.73325859     X1 -1.1169493
#> 2  -0.05355055 -0.5975271 -0.73325859     X2 -0.3797729
#> 3  -0.71706658 -1.3371989 -0.59450171     X1 -1.5187659
#> 4  -0.71706658 -1.3371989 -0.59450171     X2  0.5152054
#> 5  -0.47679959 -0.2440668 -0.23942869     X1 -1.0760903
#> 6  -0.47679959 -0.2440668 -0.23942869     X2 -0.8297431
#> 7  -1.50514840 -0.2975150  0.35554684     X1 -0.3582008
#> 8  -1.50514840 -0.2975150  0.35554684     X2 -1.6478656
#> 9  -1.97604360 -2.9386685  0.05208881     X1  0.4631693
#> 10 -1.97604360 -2.9386685  0.05208881     X2  0.4687747
#> 11 -0.16008072  0.8546154  0.32210769     X1  0.7326600
#> 12 -0.16008072  0.8546154  0.32210769     X2 -1.9148621
#> 13  0.16946714  0.3191793  0.08252579     X1  0.8765147
#> 14  0.16946714  0.3191793  0.08252579     X2  2.4096722
#> 15 -1.28277881 -1.3872892 -0.42000671     X1 -0.3812761
#> 16 -1.28277881 -1.3872892 -0.42000671     X2 -1.5716870
#> 17 -1.21252069  0.9408732 -0.91226906     X1  0.9147009
#> 18 -1.21252069  0.9408732 -0.91226906     X2  0.7508417
#> 19 -0.64700133 -0.8030641 -1.35189781     X1  0.9910317
#> 20 -0.64700133 -0.8030641 -1.35189781     X2  0.7972169