Pivot data from long to wide

pivot_wider() "widens" data, increasing the number of columns and decreasing the number of rows. The inverse transformation is pivot_longer().

pivot_wider(
  data,
  id_cols = NULL,
  values_from = "Value",
  names_from = "Name",
  names_sep = "_",
  names_prefix = "",
  names_glue = NULL,
  values_fill = NULL,
  ...
)

Arguments

data

data.frame. The data to pivot.

id_cols

character(1). The name of the column that identifies the rows. If NULL, it will use all the unique rows.

values_from

character(n). The name of the column that contains the values to be used as future variable values.

names_from

character(n). The name of the column(s) that contains the levels to be used as future column names.

names_sep

character(1). If names_from or values_from contains multiple variables, this will be used to join their values together into a single string to use as a column name.

names_prefix

character(1). String added to the start of every variable name. This is particularly useful if names_from is a numeric vector and you want to create syntactic variable names.

names_glue

character(1). Instead of names_sep and names_prefix, you can supply a glue specification that uses the names_from columns to create custom column names. Note that the only delimiters supported by names_glue are curly brackets, { and }.

values_fill

numeric(n). Optionally, a (scalar) value that will be used to replace missing values in the new columns created.

...

Not used for now.

Value

If a tibble was provided as input, pivot_wider() also returns a tibble. Otherwise, it returns a data frame.

Examples

data_long <- read.table(header = TRUE, text = "
 subject sex condition measurement
       1   M   control         7.9
       1   M     cond1        12.3
       1   M     cond2        10.7
       2   F   control         6.3
       2   F     cond1        10.6
       2   F     cond2        11.1
       3   F   control         9.5
       3   F     cond1        13.1
       3   F     cond2        13.8
       4   M   control        11.5
       4   M     cond1        13.4
       4   M     cond2        12.9")

pivot_wider(
  data_long,
  id_cols = "subject",
  names_from = "condition",
  values_from = "measurement"
)
#>   subject sex control cond1 cond2
#> 1       1   M     7.9  12.3  10.7
#> 2       2   F     6.3  10.6  11.1
#> 3       3   F     9.5  13.1  13.8
#> 4       4   M    11.5  13.4  12.9

pivot_wider(
  data_long,
  id_cols = "subject",
  names_from = "condition",
  values_from = "measurement",
  names_prefix = "Var.",
  names_sep = "."
)
#>   subject sex Var.control Var.cond1 Var.cond2
#> 1       1   M         7.9      12.3      10.7
#> 2       2   F         6.3      10.6      11.1
#> 3       3   F         9.5      13.1      13.8
#> 4       4   M        11.5      13.4      12.9

production <- expand.grid(
  product = c("A", "B"),
  country = c("AI", "EI"),
  year = 2000:2014
) %>%
  filter((product == "A" & country == "AI") | product == "B") %>%
  mutate(production = rnorm(nrow(.)))

pivot_wider(
  production,
  names_from = c("product", "country"),
  values_from = "production",
  names_glue = "prod_{product}_{country}"
)
#>    year   prod_A_AI  prod_B_AI   prod_B_EI
#> 1  2000  1.14241317 -0.3831379  0.35245599
#> 2  2001  0.03191248 -0.9820335 -0.12215535
#> 3  2002  0.52240531  1.5587061 -0.44958420
#> 4  2003  1.24272719  0.3900455  0.68505380
#> 5  2004 -1.84815547  0.7981349 -0.84238873
#> 6  2005  0.32627048  1.2697746  0.75852623
#> 7  2006  0.88379067 -0.1682324  1.31496935
#> 8  2007 -0.13617191  0.3858194 -0.48229883
#> 9  2008  0.80792145  0.5181071 -1.66020077
#> 10 2009 -0.07013075  0.9909640  0.54583622
#> 11 2010 -0.19476516 -0.6607055 -0.03719065
#> 12 2011 -0.11427203  0.2788210  0.00844293
#> 13 2012 -1.02906069 -1.4909122 -2.14526867
#> 14 2013 -0.27575917 -0.4867219 -0.70580969
#> 15 2014  0.93260485 -1.3230127  1.38936933