Select/relocate variables by name

Choose or relocate variables from a data.frame. select() keeps only the variables you mention; relocate() keeps all the variables.

relocate(.data, ..., .before = NULL, .after = NULL)

select(.data, ...)

Arguments

.data
A data.frame.

...
The name(s) of the column(s) to select.

.before, .after
Destination of the columns selected by .... Supplying neither will move the columns to the left-hand side whereas supplying both will result in an error.

Value

A data.frame.

Useful functions

There are a number of special functions which are designed to work in select() and relocate():

starts_with(), ends_with(), contains()

matches()

num_range()

everything()

Examples

select(mtcars, mpg:cyl)

#>                      mpg cyl
#> Mazda RX4           21.0   6
#> Mazda RX4 Wag       21.0   6
#> Datsun 710          22.8   4
#> Hornet 4 Drive      21.4   6
#> Hornet Sportabout   18.7   8
#> Valiant             18.1   6
#> Duster 360          14.3   8
#> Merc 240D           24.4   4
#> Merc 230            22.8   4
#> Merc 280            19.2   6
#> Merc 280C           17.8   6
#> Merc 450SE          16.4   8
#> Merc 450SL          17.3   8
#> Merc 450SLC         15.2   8
#> Cadillac Fleetwood  10.4   8
#> Lincoln Continental 10.4   8
#> Chrysler Imperial   14.7   8
#> Fiat 128            32.4   4
#> Honda Civic         30.4   4
#> Toyota Corolla      33.9   4
#> Toyota Corona       21.5   4
#> Dodge Challenger    15.5   8
#> AMC Javelin         15.2   8
#> Camaro Z28          13.3   8
#> Pontiac Firebird    19.2   8
#> Fiat X1-9           27.3   4
#> Porsche 914-2       26.0   4
#> Lotus Europa        30.4   4
#> Ford Pantera L      15.8   8
#> Ferrari Dino        19.7   6
#> Maserati Bora       15.0   8
#> Volvo 142E          21.4   4

select(mtcars, MilesPerGallon = mpg, Cylinders = cyl)

#>                     MilesPerGallon Cylinders
#> Mazda RX4                     21.0         6
#> Mazda RX4 Wag                 21.0         6
#> Datsun 710                    22.8         4
#> Hornet 4 Drive                21.4         6
#> Hornet Sportabout             18.7         8
#> Valiant                       18.1         6
#> Duster 360                    14.3         8
#> Merc 240D                     24.4         4
#> Merc 230                      22.8         4
#> Merc 280                      19.2         6
#> Merc 280C                     17.8         6
#> Merc 450SE                    16.4         8
#> Merc 450SL                    17.3         8
#> Merc 450SLC                   15.2         8
#> Cadillac Fleetwood            10.4         8
#> Lincoln Continental           10.4         8
#> Chrysler Imperial             14.7         8
#> Fiat 128                      32.4         4
#> Honda Civic                   30.4         4
#> Toyota Corolla                33.9         4
#> Toyota Corona                 21.5         4
#> Dodge Challenger              15.5         8
#> AMC Javelin                   15.2         8
#> Camaro Z28                    13.3         8
#> Pontiac Firebird              19.2         8
#> Fiat X1-9                     27.3         4
#> Porsche 914-2                 26.0         4
#> Lotus Europa                  30.4         4
#> Ford Pantera L                15.8         8
#> Ferrari Dino                  19.7         6
#> Maserati Bora                 15.0         8
#> Volvo 142E                    21.4         4

mtcars %>% select(mpg)

#>                      mpg
#> Mazda RX4           21.0
#> Mazda RX4 Wag       21.0
#> Datsun 710          22.8
#> Hornet 4 Drive      21.4
#> Hornet Sportabout   18.7
#> Valiant             18.1
#> Duster 360          14.3
#> Merc 240D           24.4
#> Merc 230            22.8
#> Merc 280            19.2
#> Merc 280C           17.8
#> Merc 450SE          16.4
#> Merc 450SL          17.3
#> Merc 450SLC         15.2
#> Cadillac Fleetwood  10.4
#> Lincoln Continental 10.4
#> Chrysler Imperial   14.7
#> Fiat 128            32.4
#> Honda Civic         30.4
#> Toyota Corolla      33.9
#> Toyota Corona       21.5
#> Dodge Challenger    15.5
#> AMC Javelin         15.2
#> Camaro Z28          13.3
#> Pontiac Firebird    19.2
#> Fiat X1-9           27.3
#> Porsche 914-2       26.0
#> Lotus Europa        30.4
#> Ford Pantera L      15.8
#> Ferrari Dino        19.7
#> Maserati Bora       15.0
#> Volvo 142E          21.4

mtcars %>% select(!mpg, !cyl)

#>                      disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128             78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic          75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla       71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9            79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa         95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          121.0 109 4.11 2.780 18.60  1  1    4    2

iris %>% select(contains("Petal"))

#>     Petal.Length Petal.Width
#> 1            1.4         0.2
#> 2            1.4         0.2
#> 3            1.3         0.2
#> 4            1.5         0.2
#> 5            1.4         0.2
#> 6            1.7         0.4
#> 7            1.4         0.3
#> 8            1.5         0.2
#> 9            1.4         0.2
#> 10           1.5         0.1
#> 11           1.5         0.2
#> 12           1.6         0.2
#> 13           1.4         0.1
#> 14           1.1         0.1
#> 15           1.2         0.2
#> 16           1.5         0.4
#> 17           1.3         0.4
#> 18           1.4         0.3
#> 19           1.7         0.3
#> 20           1.5         0.3
#> 21           1.7         0.2
#> 22           1.5         0.4
#> 23           1.0         0.2
#> 24           1.7         0.5
#> 25           1.9         0.2
#> 26           1.6         0.2
#> 27           1.6         0.4
#> 28           1.5         0.2
#> 29           1.4         0.2
#> 30           1.6         0.2
#> 31           1.6         0.2
#> 32           1.5         0.4
#> 33           1.5         0.1
#> 34           1.4         0.2
#> 35           1.5         0.2
#> 36           1.2         0.2
#> 37           1.3         0.2
#> 38           1.4         0.1
#> 39           1.3         0.2
#> 40           1.5         0.2
#> 41           1.3         0.3
#> 42           1.3         0.3
#> 43           1.3         0.2
#> 44           1.6         0.6
#> 45           1.9         0.4
#> 46           1.4         0.3
#> 47           1.6         0.2
#> 48           1.4         0.2
#> 49           1.5         0.2
#> 50           1.4         0.2
#> 51           4.7         1.4
#> 52           4.5         1.5
#> 53           4.9         1.5
#> 54           4.0         1.3
#> 55           4.6         1.5
#> 56           4.5         1.3
#> 57           4.7         1.6
#> 58           3.3         1.0
#> 59           4.6         1.3
#> 60           3.9         1.4
#> 61           3.5         1.0
#> 62           4.2         1.5
#> 63           4.0         1.0
#> 64           4.7         1.4
#> 65           3.6         1.3
#> 66           4.4         1.4
#> 67           4.5         1.5
#> 68           4.1         1.0
#> 69           4.5         1.5
#> 70           3.9         1.1
#> 71           4.8         1.8
#> 72           4.0         1.3
#> 73           4.9         1.5
#> 74           4.7         1.2
#> 75           4.3         1.3
#> 76           4.4         1.4
#> 77           4.8         1.4
#> 78           5.0         1.7
#> 79           4.5         1.5
#> 80           3.5         1.0
#> 81           3.8         1.1
#> 82           3.7         1.0
#> 83           3.9         1.2
#> 84           5.1         1.6
#> 85           4.5         1.5
#> 86           4.5         1.6
#> 87           4.7         1.5
#> 88           4.4         1.3
#> 89           4.1         1.3
#> 90           4.0         1.3
#> 91           4.4         1.2
#> 92           4.6         1.4
#> 93           4.0         1.2
#> 94           3.3         1.0
#> 95           4.2         1.3
#> 96           4.2         1.2
#> 97           4.2         1.3
#> 98           4.3         1.3
#> 99           3.0         1.1
#> 100          4.1         1.3
#> 101          6.0         2.5
#> 102          5.1         1.9
#> 103          5.9         2.1
#> 104          5.6         1.8
#> 105          5.8         2.2
#> 106          6.6         2.1
#> 107          4.5         1.7
#> 108          6.3         1.8
#> 109          5.8         1.8
#> 110          6.1         2.5
#> 111          5.1         2.0
#> 112          5.3         1.9
#> 113          5.5         2.1
#> 114          5.0         2.0
#> 115          5.1         2.4
#> 116          5.3         2.3
#> 117          5.5         1.8
#> 118          6.7         2.2
#> 119          6.9         2.3
#> 120          5.0         1.5
#> 121          5.7         2.3
#> 122          4.9         2.0
#> 123          6.7         2.0
#> 124          4.9         1.8
#> 125          5.7         2.1
#> 126          6.0         1.8
#> 127          4.8         1.8
#> 128          4.9         1.8
#> 129          5.6         2.1
#> 130          5.8         1.6
#> 131          6.1         1.9
#> 132          6.4         2.0
#> 133          5.6         2.2
#> 134          5.1         1.5
#> 135          5.6         1.4
#> 136          6.1         2.3
#> 137          5.6         2.4
#> 138          5.5         1.8
#> 139          4.8         1.8
#> 140          5.4         2.1
#> 141          5.6         2.4
#> 142          5.1         2.3
#> 143          5.1         1.9
#> 144          5.9         2.3
#> 145          5.7         2.5
#> 146          5.2         2.3
#> 147          5.0         1.9
#> 148          5.2         2.0
#> 149          5.4         2.3
#> 150          5.1         1.8

df <- as.data.frame(matrix(runif(100), nrow = 10))
df <- as.data.frame(df[c(3, 4, 7, 1, 9, 8, 5, 2, 6, 10)])
df %>% select(num_range("V", 4:6))

#>            V4        V5        V6
#> 1  0.16218885 0.1656501 0.1364388
#> 2  0.45478598 0.6954764 0.9997514
#> 3  0.76041597 0.8675843 0.1072434
#> 4  0.16309973 0.3624767 0.3182820
#> 5  0.21895608 0.0909190 0.2243881
#> 6  0.51703127 0.3853781 0.4505799
#> 7  0.76511371 0.9325069 0.4506843
#> 8  0.01241407 0.8551653 0.7905153
#> 9  0.26255770 0.7432032 0.2441034
#> 10 0.04697012 0.8253677 0.6516112

mtcars %>% relocate(ends_with("p"), .before = mpg)

#>                      disp  hp  mpg cyl drat    wt  qsec vs am gear carb
#> Mazda RX4           160.0 110 21.0   6 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       160.0 110 21.0   6 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          108.0  93 22.8   4 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      258.0 110 21.4   6 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   360.0 175 18.7   8 3.15 3.440 17.02  0  0    3    2
#> Valiant             225.0 105 18.1   6 2.76 3.460 20.22  1  0    3    1
#> Duster 360          360.0 245 14.3   8 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           146.7  62 24.4   4 3.69 3.190 20.00  1  0    4    2
#> Merc 230            140.8  95 22.8   4 3.92 3.150 22.90  1  0    4    2
#> Merc 280            167.6 123 19.2   6 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           167.6 123 17.8   6 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          275.8 180 16.4   8 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          275.8 180 17.3   8 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         275.8 180 15.2   8 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  472.0 205 10.4   8 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 460.0 215 10.4   8 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   440.0 230 14.7   8 3.23 5.345 17.42  0  0    3    4
#> Fiat 128             78.7  66 32.4   4 4.08 2.200 19.47  1  1    4    1
#> Honda Civic          75.7  52 30.4   4 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla       71.1  65 33.9   4 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       120.1  97 21.5   4 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    318.0 150 15.5   8 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         304.0 150 15.2   8 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          350.0 245 13.3   8 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    400.0 175 19.2   8 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9            79.0  66 27.3   4 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       120.3  91 26.0   4 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa         95.1 113 30.4   4 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      351.0 264 15.8   8 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        145.0 175 19.7   6 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       301.0 335 15.0   8 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          121.0 109 21.4   4 4.11 2.780 18.60  1  1    4    2