Perform an operation with temporary groups

This function allows you to modify the grouping variables for a single operation.

with_groups(.data, .groups, .f, ...)

Arguments

.data

A data.frame.

.groups

<poor-select> One or more variables to group by. Unlike group_by(), you can only group by existing variables, and you can use poor-select syntax like c(x, y, z) to select multiple variables.

Use NULL to temporarily ungroup.

.f

A function to apply to regrouped data. Supports lambda-style ~ syntax.

...

Additional arguments passed on to .f.

Examples

df <- data.frame(g = c(1, 1, 2, 2, 3), x = runif(5))
df %>% with_groups(g, mutate, x_mean = mean(x))
#>   g         x    x_mean
#> 1 1 0.4087823 0.2958394
#> 2 1 0.1828964 0.2958394
#> 3 2 0.6715215 0.4391016
#> 4 2 0.2066817 0.4391016
#> 5 3 0.4538612 0.4538612
df %>% with_groups(g, ~ mutate(.x, x_mean = mean(x)))
#>   g         x    x_mean
#> 1 1 0.4087823 0.2958394
#> 2 1 0.1828964 0.2958394
#> 3 2 0.6715215 0.4391016
#> 4 2 0.2066817 0.4391016
#> 5 3 0.4538612 0.4538612

df %>%
  group_by(g) %>%
  with_groups(NULL, mutate, x_mean = mean(x))
#>   g         x    x_mean
#> 1 1 0.4087823 0.3847486
#> 2 1 0.1828964 0.3847486
#> 3 2 0.6715215 0.3847486
#> 4 2 0.2066817 0.3847486
#> 5 3 0.4538612 0.3847486

# NB: grouping can't be restored if you remove the grouping variables
df %>%
  group_by(g) %>%
  with_groups(NULL, mutate, g = NULL)
#>           x
#> 1 0.4087823
#> 2 0.1828964
#> 3 0.6715215
#> 4 0.2066817
#> 5 0.4538612