Context dependent expressions

These functions return information about the "current" group or "current" variable, so only work inside specific contexts like summarise() and mutate().

n() gives the number of observations in the current group.

cur_data() gives the current data for the current group (excluding grouping variables).

cur_group() gives the group keys, a single row data.frame containing a column for each grouping variable and its value.

cur_group_id() gives a unique numeric identifier for the current group.

cur_group_rows() gives the rows the groups appear in the data.

n()

cur_data()

cur_group()

cur_group_id()

cur_group_rows()

data.table

If you're familiar with data.table:

cur_data() <-> .SD

cur_group_id() <-> .GRP

cur_group() <-> .BY

cur_group_rows() <-> .I

See also

Examples

df <- data.frame(
  g = sample(rep(letters[1:3], 1:3)),
  x = runif(6),
  y = runif(6),
  stringsAsFactors = FALSE
)
gf <- df %>% group_by(g)

gf %>% summarise(n = n())

#>   g n
#> 1 a 1
#> 2 b 2
#> 3 c 3

gf %>% mutate(id = cur_group_id())

#>   g         x          y id
#> 1 c 0.9835254 0.13807484  3
#> 2 c 0.5898376 0.08084496  3
#> 3 c 0.7591584 0.65598263  3
#> 4 b 0.8360753 0.60200386  2
#> 5 b 0.7628195 0.65699583  2
#> 6 a 0.4172699 0.32931716  1

gf %>% summarise(row = cur_group_rows())

#>   g row
#> 1 a   6
#> 2 b   4
#> 3 b   5
#> 4 c   1
#> 5 c   2
#> 6 c   3

gf %>% summarise(data = list(cur_group()))

#>   g data
#> 1 a    a
#> 2 b    b
#> 3 c    c

gf %>% summarise(data = list(cur_data()))

#>   g         data
#> 1 a 0.417269....
#> 2 b c(0.8360....
#> 3 c c(0.9835....