Context dependent expressions

These functions return information about the "current" group or "current" variable, so only work inside specific contexts like summarise() and mutate().

n() gives the number of observations in the current group.

cur_data() gives the current data for the current group (excluding grouping variables).

cur_data_all() gives the current data for the current group (including grouping variables).

cur_group() gives the group keys, a single row data.frame containing a column for each grouping variable and its value.

cur_group_id() gives a unique numeric identifier for the current group.

cur_group_rows() gives the rows the groups appear in the data.

cur_column() gives the name of the current column (in across() only).

Usage

n()

cur_data()

cur_data_all()

cur_group()

cur_group_id()

cur_group_rows()

cur_column()

data.table

If you're familiar with data.table:

cur_data() <-> .SD

cur_group_id() <-> .GRP

cur_group() <-> .BY

cur_group_rows() <-> .I

See also

Examples

df <- data.frame(
  g = sample(rep(letters[1:3], 1:3)),
  x = runif(6),
  y = runif(6),
  stringsAsFactors = FALSE
)
gf <- df %>% group_by(g)

gf %>% summarise(n = n())
#>   g n
#> 1 a 1
#> 2 b 2
#> 3 c 3

gf %>% mutate(id = cur_group_id())
#>   g         x         y id
#> 1 c 0.8807781 0.8725876  3
#> 2 c 0.2902445 0.8048914  3
#> 3 b 0.5364767 0.1552964  2
#> 4 a 0.6629266 0.4571946  1
#> 5 c 0.3309674 0.2367926  3
#> 6 b 0.5720669 0.8768752  2
gf %>% summarise(row = cur_group_rows())
#> `summarise()` has grouped output by 'g'. You can override using the `.groups` argument.
#>   g row
#> 1 a   4
#> 2 b   3
#> 3 b   6
#> 4 c   1
#> 5 c   2
#> 6 c   5
gf %>% summarise(data = list(cur_group()))
#>   g data
#> 1 a    a
#> 2 b    b
#> 3 c    c
gf %>% summarise(data = list(cur_data()))
#>   g                                                             data
#> 1 a                                             0.6629266, 0.4571946
#> 2 b                       0.5364767, 0.5720669, 0.1552964, 0.8768752
#> 3 c 0.8807781, 0.2902445, 0.3309674, 0.8725876, 0.8048914, 0.2367926
gf %>% summarise(data = list(cur_data_all()))
#>   g
#> 1 a
#> 2 b
#> 3 c
#>                                                                                                                        data
#> 1                                                                                   a, 0.662926638033241, 0.457194555550814
#> 2                                          b, b, 0.536476703826338, 0.572066916385666, 0.155296449549496, 0.876875170739368
#> 3 c, c, c, 0.880778069375083, 0.290244481759146, 0.330967405345291, 0.872587644029409, 0.804891440784559, 0.236792633775622

gf %>% mutate(across(everything(), ~ paste(cur_column(), round(.x, 2))))
#>   g      x      y
#> 1 c x 0.88 y 0.87
#> 2 c x 0.29  y 0.8
#> 3 b x 0.54 y 0.16
#> 4 a x 0.66 y 0.46
#> 5 c x 0.33 y 0.24
#> 6 b x 0.57 y 0.88