Subset distinct/unique rows

Select only distinct/unique rows from a data.frame.

Usage

distinct(.data, ..., .keep_all = FALSE)

Arguments

.data

A data.frame.

...

Optional variables to use when determining uniqueness. If there are multiple rows for a given combination of inputs, only the first row will be preserved. If omitted, will use all variables.

.keep_all

logical(1). If TRUE, keep all variables in .data. If a combination of ... is not distinct, this keeps the first row of values.

Value

A data.frame with the following properties:

Rows are a subset of the input but appear in the same order.

Columns are not modified if ... is empty or .keep_all is TRUE. Otherwise, distinct() first calls mutate() to create new columns.

Groups are not modified.

data.frame attributes are preserved.

Examples

df <- data.frame(
  x = sample(10, 100, rep = TRUE),
  y = sample(10, 100, rep = TRUE)
)
nrow(df)
#> [1] 100
nrow(distinct(df))
#> [1] 58
nrow(distinct(df, x, y))
#> [1] 58

distinct(df, x)
#>     x
#> 1   2
#> 2   9
#> 3  10
#> 4   8
#> 5   5
#> 6   4
#> 7   7
#> 8   1
#> 9   3
#> 10  6
distinct(df, y)
#>     y
#> 1   6
#> 2   9
#> 3   7
#> 4   4
#> 5   3
#> 6   8
#> 7  10
#> 8   2
#> 9   5
#> 10  1

# You can choose to keep all other variables as well
distinct(df, x, .keep_all = TRUE)
#>     x  y
#> 1   2  6
#> 2   9  9
#> 3  10  9
#> 4   8  4
#> 5   5  8
#> 6   4 10
#> 7   7  9
#> 8   1  9
#> 9   3  3
#> 10  6  3
distinct(df, y, .keep_all = TRUE)
#>     x  y
#> 1   2  6
#> 2   9  9
#> 3   9  7
#> 4   8  4
#> 5  10  3
#> 6   5  8
#> 7   4 10
#> 8   4  2
#> 9   5  5
#> 10  4  1

# You can also use distinct on computed variables
distinct(df, diff = abs(x - y))
#>   diff
#> 1    4
#> 2    0
#> 3    1
#> 4    2
#> 5    7
#> 6    3
#> 7    6
#> 8    8
#> 9    5

# The same behaviour applies for grouped data frames,
# except that the grouping variables are always included
df <- data.frame(
  g = c(1, 1, 2, 2),
  x = c(1, 1, 2, 1)
) %>% group_by(g)
df %>% distinct(x)
#>   g x
#> 1 1 1
#> 2 2 2
#> 3 2 1