Subset distinct/unique rows

Select only distinct/unique rows from a data.frame.

distinct(.data, ..., .keep_all = FALSE)

Arguments

.data

A data.frame.

...

Optional variables to use when determining uniqueness. If there are multiple rows for a given combination of inputs, only the first row will be preserved. If omitted, will use all variables.

.keep_all

logical(1). If TRUE, keep all variables in .data. If a combination of ... is not distinct, this keeps the first row of values.

Value

A data.frame with the following properties:

Rows are a subset of the input but appear in the same order.

Columns are not modified if ... is empty or .keep_all is TRUE. Otherwise, distinct() first calls mutate() to create new columns.

Groups are not modified.

data.frame attributes are preserved.

Examples

df <- data.frame(
  x = sample(10, 100, rep = TRUE),
  y = sample(10, 100, rep = TRUE)
)
nrow(df)
#> [1] 100
nrow(distinct(df))
#> [1] 59
nrow(distinct(df, x, y))
#> [1] 59

distinct(df, x)
#>     x
#> 1   5
#> 2  10
#> 3   8
#> 4   2
#> 5   4
#> 6   1
#> 7   7
#> 8   9
#> 9   6
#> 10  3
distinct(df, y)
#>     y
#> 1   9
#> 2   8
#> 3   5
#> 4   3
#> 5  10
#> 6   7
#> 7   6
#> 8   2
#> 9   1
#> 10  4

# You can choose to keep all other variables as well
distinct(df, x, .keep_all = TRUE)
#>     x  y
#> 1   5  9
#> 2  10  5
#> 3   8  3
#> 4   2 10
#> 5   4  7
#> 6   1  7
#> 7   7  9
#> 8   9  1
#> 9   6  5
#> 10  3  4
distinct(df, y, .keep_all = TRUE)
#>     x  y
#> 1   5  9
#> 2   5  8
#> 3  10  5
#> 4   8  3
#> 5   2 10
#> 6   4  7
#> 7   1  6
#> 8   8  2
#> 9   9  1
#> 10  8  4

# You can also use distinct on computed variables
distinct(df, diff = abs(x - y))
#>    diff
#> 1     4
#> 2     3
#> 3     5
#> 4     8
#> 5     6
#> 6     2
#> 7     7
#> 8     1
#> 9     9
#> 10    0

# The same behaviour applies for grouped data frames,
# except that the grouping variables are always included
df <- data.frame(
  g = c(1, 1, 2, 2),
  x = c(1, 1, 2, 1)
) %>% group_by(g)
df %>% distinct(x)
#>   g x
#> 1 1 1
#> 2 2 2
#> 3 2 1