Pivot data from wide to long

pivot_longer() "lengthens" data, increasing the number of rows and decreasing the number of columns. The inverse transformation is pivot_wider().

Usage

pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  ...
)

Arguments

data

data.frame. The data to pivot.

cols

<poor-select>. Columns to pivot into longer format.

names_to

character(n). The name of the new column(s) that will contain the column names.

names_prefix

character(1). A regular expression used to remove matching text from the start of each variable name.

names_sep, names_pattern

character(1). If names_to contains multiple values, this argument controls how the column name is broken up. names_pattern takes a regular expression containing matching groups (()⁠).

values_to

character(n). The name of the new column(s) that will contain the values of the pivoted variables.

values_drop_na

logical(1). If TRUE, will drop rows that contain only NA in the values_to column. This effectively converts explicit missing values to implicit missing values, and should generally be used only when missing values in data were created by its structure.

...

Additional arguments passed on to methods.

Value

A data.frame.

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))
# Customizing the names
pivot_longer(
  data = wide_data,
  cols = c(1, 2),
  names_to = "Column",
  values_to = "Numbers"
)
#>            X3          X4         X5 Column     Numbers
#> 1  -0.6776533 -0.73990353  2.7847301     X1 -0.05747513
#> 2  -0.6776533 -0.73990353  2.7847301     X2 -0.45416870
#> 3   0.3637399  0.28061078  0.5065914     X1  0.96190111
#> 4   0.3637399  0.28061078  0.5065914     X2 -1.56305060
#> 5  -0.8312748 -1.99578763 -0.8503908     X1  0.29790540
#> 6  -0.8312748 -1.99578763 -0.8503908     X2  0.37035124
#> 7   0.5760886  0.84975331 -1.3899269     X1 -0.61494612
#> 8   0.5760886  0.84975331 -1.3899269     X2  0.43348685
#> 9   0.1115160  0.57816236  0.3872362     X1  0.56357121
#> 10  0.1115160  0.57816236  0.3872362     X2 -0.07061861
#> 11  0.6587789 -0.25511126 -0.7198510     X1 -0.54522059
#> 12  0.6587789 -0.25511126 -0.7198510     X2  1.19722239
#> 13  0.4855238  0.21081904  0.6195347     X1  1.07414876
#> 14  0.4855238  0.21081904  0.6195347     X2 -1.70626994
#> 15 -1.4221677  2.60356907 -0.4884581     X1  0.42308401
#> 16 -1.4221677  2.60356907 -0.4884581     X2 -0.32856831
#> 17  1.7800330  0.17331458  2.5781728     X1  1.97342742
#> 18  1.7800330  0.17331458  2.5781728     X2  0.60757738
#> 19 -1.8754422 -0.02527595 -1.0380768     X1 -1.24833687
#> 20 -1.8754422 -0.02527595 -1.0380768     X2 -0.08940164