Pivot data from wide to long

pivot_longer() "lengthens" data, increasing the number of rows and decreasing the number of columns. The inverse transformation is pivot_wider().

Usage

pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  ...
)

Arguments

data

data.frame. The data to pivot.

cols

<poor-select>. Columns to pivot into longer format.

names_to

character(n). The name of the new column(s) that will contain the column names.

names_prefix

character(1). A regular expression used to remove matching text from the start of each variable name.

names_sep, names_pattern

character(1). If names_to contains multiple values, this argument controls how the column name is broken up. names_pattern takes a regular expression containing matching groups (()).

values_to

character(n). The name of the new column(s) that will contain the values of the pivoted variables.

values_drop_na

logical(1). If TRUE, will drop rows that contain only NA in the values_to column. This effectively converts explicit missing values to implicit missing values, and should generally be used only when missing values in data were created by its structure.

...

Additional arguments passed on to methods.

Value

A data.frame.

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))
# Customizing the names
pivot_longer(
  data = wide_data,
  cols = c(1, 2),
  names_to = "Column",
  values_to = "Numbers"
)
#>             X3         X4         X5 Column    Numbers
#> 1  -0.46134003 -1.5699950 -0.9748966     X1 -0.4157757
#> 2  -0.46134003 -1.5699950 -0.9748966     X2  1.4798022
#> 3   0.96065190 -0.2927193 -0.2696236     X1  0.8783356
#> 4   0.96065190 -0.2927193 -0.2696236     X2 -0.9220620
#> 5   0.06562962  0.7015996 -0.5358896     X1 -0.3517207
#> 6   0.06562962  0.7015996 -0.5358896     X2 -0.8835623
#> 7   0.55487373  0.3818974 -0.7060552     X1  1.3601535
#> 8   0.55487373  0.3818974 -0.7060552     X2  1.7803711
#> 9  -1.15584374  0.4512109  0.4710315     X1  0.5795432
#> 10 -1.15584374  0.4512109  0.4710315     X2 -0.7743281
#> 11 -1.03720294  0.7212278  1.8307889     X1 -1.3427687
#> 12 -1.03720294  0.7212278  1.8307889     X2  0.1919454
#> 13 -0.42198560 -0.9044911  1.1859084     X1  2.0168895
#> 14 -0.42198560 -0.9044911  1.1859084     X2  0.8433395
#> 15  0.82655970 -1.2482573 -0.8592202     X1  0.4671025
#> 16  0.82655970 -1.2482573 -0.8592202     X2 -0.4366906
#> 17 -0.07363882  0.9322790  1.0148193     X1  0.2937089
#> 18 -0.07363882  0.9322790  1.0148193     X2  1.9686142
#> 19 -1.05880805 -1.2035600 -1.1869582     X1  0.1270592
#> 20 -1.05880805 -1.2035600 -1.1869582     X2 -0.3997362