Pivot data from wide to long

pivot_longer() "lengthens" data, increasing the number of rows and decreasing the number of columns. The inverse transformation is pivot_wider().

pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  ...
)

Arguments

data

data.frame. The data to pivot.

cols

<poor-select>. Columns to pivot into longer format.

names_to

character(n). The name of the new column(s) that will contain the column names.

names_prefix

character(1). A regular expression used to remove matching text from the start of each variable name.

names_sep, names_pattern

character(1). If names_to contains multiple values, this argument controls how the column name is broken up. names_pattern takes a regular expression containing matching groups (()⁠).

values_to

character(n). The name of the new column(s) that will contain the values of the pivoted variables.

values_drop_na

logical(1). If TRUE, will drop rows that contain only NA in the values_to column. This effectively converts explicit missing values to implicit missing values, and should generally be used only when missing values in data were created by its structure.

...

Additional arguments passed on to methods.

Value

A data.frame.

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))
# Customizing the names
pivot_longer(
  data = wide_data,
  cols = c(1, 2),
  names_to = "Column",
  values_to = "Numbers"
)
#>              X3          X4         X5 Column     Numbers
#> 1  -1.177028384 -1.12556434  0.3081741     X1  0.92507563
#> 2  -1.177028384 -1.12556434  0.3081741     X2  0.63122310
#> 3  -0.008365366 -0.03267294  1.2474795     X1 -0.21808690
#> 4  -0.008365366 -0.03267294  1.2474795     X2  0.09843901
#> 5  -0.642591829 -0.62665808 -1.6741486     X1 -1.69402483
#> 6  -0.642591829 -0.62665808 -1.6741486     X2 -0.23550479
#> 7  -0.663629566  0.42811261  0.7728313     X1  0.76780395
#> 8  -0.663629566  0.42811261  0.7728313     X2  0.20034282
#> 9  -0.193925517 -0.57231652  1.6156469     X1 -2.70992926
#> 10 -0.193925517 -0.57231652  1.6156469     X2  0.00214565
#> 11  0.856010209  0.72641917 -0.7537741     X1 -0.48327025
#> 12  0.856010209  0.72641917 -0.7537741     X2  0.28499498
#> 13 -0.159096915 -0.13889225 -0.6942943     X1  1.09479010
#> 14 -0.159096915 -0.13889225 -0.6942943     X2  1.22825035
#> 15  1.285588382  0.16151977  0.4922455     X1  1.18017520
#> 16  1.285588382  0.16151977  0.4922455     X2 -0.13682887
#> 17  0.951302742 -0.65619343 -1.8394339     X1 -0.08303652
#> 18  0.951302742 -0.65619343 -1.8394339     X2 -0.23591306
#> 19 -0.046217781  0.52106858 -0.2374241     X1  0.91001123
#> 20 -0.046217781  0.52106858 -0.2374241     X2  1.04937681