Pivot data from wide to long

pivot_longer() "lengthens" data, increasing the number of rows and decreasing the number of columns. The inverse transformation is pivot_wider().

Usage

pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  ...
)

Arguments

data

data.frame. The data to pivot.

cols

<poor-select>. Columns to pivot into longer format.

names_to

character(n). The name of the new column(s) that will contain the column names.

names_prefix

character(1). A regular expression used to remove matching text from the start of each variable name.

names_sep, names_pattern

character(1). If names_to contains multiple values, this argument controls how the column name is broken up. names_pattern takes a regular expression containing matching groups (()⁠).

values_to

character(n). The name of the new column(s) that will contain the values of the pivoted variables.

values_drop_na

logical(1). If TRUE, will drop rows that contain only NA in the values_to column. This effectively converts explicit missing values to implicit missing values, and should generally be used only when missing values in data were created by its structure.

...

Additional arguments passed on to methods.

Value

A data.frame.

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))
# Customizing the names
pivot_longer(
  data = wide_data,
  cols = c(1, 2),
  names_to = "Column",
  values_to = "Numbers"
)
#>            X3          X4          X5 Column    Numbers
#> 1   1.3498533  0.02055760  0.09443817     X1 -0.2332569
#> 2   1.3498533  0.02055760  0.09443817     X2 -1.8911124
#> 3   0.1538122 -1.93920367 -1.57572418     X1 -1.0584832
#> 4   0.1538122 -1.93920367 -1.57572418     X2 -1.2596764
#> 5   1.4577890 -0.02754891  1.11810792     X1 -0.4944809
#> 6   1.4577890 -0.02754891  1.11810792     X2 -0.6856535
#> 7   1.0270109  0.28359821  0.26748160     X1  1.0074206
#> 8   1.0270109  0.28359821  0.26748160     X2  2.2519525
#> 9  -2.3795247 -0.01450952 -1.25200254     X1  0.8384102
#> 10 -2.3795247 -0.01450952 -1.25200254     X2 -1.7464295
#> 11 -1.5538148 -0.12761286 -1.43745323     X1 -1.1180250
#> 12 -1.5538148 -0.12761286 -1.43745323     X2  0.9425553
#> 13 -1.5064117  3.23352115 -0.65373941     X1 -0.5691817
#> 14 -1.5064117  3.23352115 -0.65373941     X2 -0.5969161
#> 15  1.1593219 -0.76652508  0.55710956     X1 -0.2059145
#> 16  1.1593219 -0.76652508  0.55710956     X2 -0.5523718
#> 17 -0.1538383 -0.89023229  2.19479275     X1  0.3907815
#> 18 -0.1538383 -0.89023229  2.19479275     X2 -1.4218948
#> 19 -1.3903249 -0.78468457 -1.53679326     X1 -1.5621200
#> 20 -1.3903249 -0.78468457 -1.53679326     X2  1.1021860