Pivot data from wide to long

pivot_longer() "lengthens" data, increasing the number of rows and decreasing the number of columns. The inverse transformation is pivot_wider().

Usage

pivot_longer(
  data,
  cols,
  names_to = "name",
  names_prefix = NULL,
  names_sep = NULL,
  names_pattern = NULL,
  values_to = "value",
  values_drop_na = FALSE,
  ...
)

Arguments

data

data.frame. The data to pivot.

cols

<poor-select>. Columns to pivot into longer format.

names_to

character(n). The name of the new column(s) that will contain the column names.

names_prefix

character(1). A regular expression used to remove matching text from the start of each variable name.

names_sep, names_pattern

character(1). If names_to contains multiple values, this argument controls how the column name is broken up. names_pattern takes a regular expression containing matching groups (()).

values_to

character(n). The name of the new column(s) that will contain the values of the pivoted variables.

values_drop_na

logical(1). If TRUE, will drop rows that contain only NA in the values_to column. This effectively converts explicit missing values to implicit missing values, and should generally be used only when missing values in data were created by its structure.

...

Additional arguments passed on to methods.

Value

A data.frame.

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))
# Customizing the names
pivot_longer(
  data = wide_data,
  cols = c(1, 2),
  names_to = "Column",
  values_to = "Numbers"
)
#>            X3         X4          X5 Column     Numbers
#> 1   1.6734479  0.2770119  1.77204192     X1 -1.33104949
#> 2   1.6734479  0.2770119  1.77204192     X2 -1.15477558
#> 3   0.8620044 -0.9063744 -0.06912362     X1  2.07353236
#> 4   0.8620044 -0.9063744 -0.06912362     X2  1.26495958
#> 5  -0.4090102  0.9613471 -0.17325818     X1  1.54275294
#> 6  -0.4090102  0.9613471 -0.17325818     X2  0.01139603
#> 7   0.9659927 -1.5767528 -0.45741530     X1 -1.23213459
#> 8   0.9659927 -1.5767528 -0.45741530     X2  0.42347340
#> 9   0.4936819  0.5219768  0.18809465     X1 -0.43293566
#> 10  0.4936819  0.5219768  0.18809465     X2  1.11985344
#> 11 -0.9858090 -1.5301141  0.66104702     X1 -1.91147262
#> 12 -0.9858090 -1.5301141  0.66104702     X2 -0.47864076
#> 13  0.4293605 -0.5184022 -0.45087924     X1  1.30774394
#> 14  0.4293605 -0.5184022 -0.45087924     X2  1.39712825
#> 15 -0.6240340 -0.6129894  0.17012756     X1 -1.32864036
#> 16 -0.6240340 -0.6129894  0.17012756     X2  1.25200612
#> 17 -1.9551294 -2.7450434  0.46578268     X1  0.59697394
#> 18 -1.9551294 -2.7450434  0.46578268     X2 -0.95583090
#> 19  0.7910036 -0.4539674  1.29311883     X1 -0.88702130
#> 20  0.7910036 -0.4539674  1.29311883     X2 -0.52904007