Pivot data from long to wide

pivot_wider() "widens" data, increasing the number of columns and decreasing the number of rows. The inverse transformation is pivot_longer().

Usage

pivot_wider(
  data,
  id_cols = NULL,
  values_from = "Value",
  names_from = "Name",
  names_sep = "_",
  names_prefix = "",
  names_glue = NULL,
  values_fill = NULL,
  ...
)

Arguments

data

data.frame. The data to pivot.

id_cols

character(1). The name of the column that identifies the rows. If NULL, it will use all the unique rows.

values_from

character(n). The name of the column that contains the values to be used as future variable values.

names_from

character(n). The name of the column(s) that contains the levels to be used as future column names.

names_sep

character(1). If names_from or values_from contains multiple variables, this will be used to join their values together into a single string to use as a column name.

names_prefix

character(1). String added to the start of every variable name. This is particularly useful if names_from is a numeric vector and you want to create syntactic variable names.

names_glue

character(1). Instead of names_sep and names_prefix, you can supply a glue specification that uses the names_from columns to create custom column names. Note that the only delimiters supported by names_glue are curly brackets, { and }.

values_fill

numeric(n). Optionally, a (scalar) value that will be used to replace missing values in the new columns created.

...

Not used for now.

Value

If a tibble was provided as input, pivot_wider() also returns a tibble. Otherwise, it returns a data frame.

Examples

data_long <- read.table(header = TRUE, text = "
 subject sex condition measurement
       1   M   control         7.9
       1   M     cond1        12.3
       1   M     cond2        10.7
       2   F   control         6.3
       2   F     cond1        10.6
       2   F     cond2        11.1
       3   F   control         9.5
       3   F     cond1        13.1
       3   F     cond2        13.8
       4   M   control        11.5
       4   M     cond1        13.4
       4   M     cond2        12.9")

pivot_wider(
  data_long,
  id_cols = "subject",
  names_from = "condition",
  values_from = "measurement"
)
#>   subject sex control cond1 cond2
#> 1       1   M     7.9  12.3  10.7
#> 2       2   F     6.3  10.6  11.1
#> 3       3   F     9.5  13.1  13.8
#> 4       4   M    11.5  13.4  12.9

pivot_wider(
  data_long,
  id_cols = "subject",
  names_from = "condition",
  values_from = "measurement",
  names_prefix = "Var.",
  names_sep = "."
)
#>   subject sex Var.control Var.cond1 Var.cond2
#> 1       1   M         7.9      12.3      10.7
#> 2       2   F         6.3      10.6      11.1
#> 3       3   F         9.5      13.1      13.8
#> 4       4   M        11.5      13.4      12.9

production <- expand.grid(
  product = c("A", "B"),
  country = c("AI", "EI"),
  year = 2000:2014
) %>%
  filter((product == "A" & country == "AI") | product == "B") %>%
  mutate(production = rnorm(nrow(.)))

pivot_wider(
  production,
  names_from = c("product", "country"),
  values_from = "production",
  names_glue = "prod_{product}_{country}"
)
#>    year  prod_A_AI    prod_B_AI  prod_B_EI
#> 1  2000 -1.0127480 -0.650247105 -1.6877956
#> 2  2001 -0.7176856 -2.853050304  0.6410721
#> 3  2002 -0.9155220  0.218076966 -1.7118081
#> 4  2003  0.0848408  0.580298123  0.2530245
#> 5  2004 -0.9519616 -1.228911977 -1.5859325
#> 6  2005  0.6447078  1.213110888 -0.1012849
#> 7  2006  0.2841706 -0.004469214  1.3169864
#> 8  2007 -0.7762699  0.412190609  0.3755996
#> 9  2008  1.2637516  0.292233195  0.6024007
#> 10 2009  0.2385699 -0.112629298  0.9517526
#> 11 2010  0.7008233 -0.215909600 -2.4774234
#> 12 2011  1.4980219  1.424975932 -1.1089933
#> 13 2012  1.3034068 -0.349897612 -0.1842110
#> 14 2013  0.4170443 -0.660588417  0.5584906
#> 15 2014 -0.1296331  0.038526992  0.8304355